editosta: move each function into the appropriate file
