+++ editoast/src/modelsv2/infra/route_from_waypoint_result.rs
@@ -1,5 +1,12 @@
+use diesel::sql_query;
+use diesel::sql_types::BigInt;
 use diesel::sql_types::Bool;
 use diesel::sql_types::Text;
+use diesel_async::RunQueryDsl;
+
+use super::Infra;
+use crate::error::Result;
+use crate::modelsv2::DbConnection;
 
 #[derive(QueryableByName)]
 pub struct RouteFromWaypointResult {
@@ -7,4 +14,21 @@ pub struct RouteFromWaypointResult {
     pub route_id: String,
     #[diesel(sql_type = Bool)]
     pub is_entry_point: bool,
+}
+
+impl Infra {
+    pub async fn get_routes_from_waypoint(
+        &self,
+        conn: &mut DbConnection,
+        waypoint_id: &String,
+        waypoint_type: String,
+    ) -> Result<Vec<RouteFromWaypointResult>> {
+        let routes = sql_query(include_str!("sql/get_routes_from_waypoint.sql"))
+            .bind::<BigInt, _>(self.id)
+            .bind::<Text, _>(&waypoint_id)
+            .bind::<Text, _>(waypoint_type)
+            .load::<RouteFromWaypointResult>(conn)
+            .await?;
+        Ok(routes)
+    }
 }
